+++ packages/angular/src/lib/phone-input/index.ts
@@ -1,0 +1,2 @@
+export type { TelixonPhoneInputOptions } from './models';
+export { TelixonPhoneInput } from './telixon-phone-input';