+++ packages/angular/src/test-setup.ts
@@ -1,0 +1,4 @@
+import { setupTestBed } from '@analogjs/vitest-angular/setup-testbed';
+import '@angular/compiler';
+
+setupTestBed({ zoneless: true, errorOnUnknownElements: true, errorOnUnknownProperties: true });